cliche/#2: added tab menu on search
