+++ src/pages/search/SearchScreen.tsx
@@ -1,12 +1,51 @@
 import Box from "@mui/material/Box";
-import { SearchHeader } from "./component/SearchHeader";
+import { SearchHeader } from "./component/header/SearchHeader";
 import { SearchResultList } from "./component/SearchResultList";
+import { useState, useMemo } from "react";
+import { SearchTabs } from "./component/SearchTabs";
+
+enum SearchFilter {
+  title = "title",
+  category = "category",
+  users = "users",
+  lists = "lists",
+}
+
+const options = [
+  { label: "Títulos", tabType: SearchFilter.title },
+  { label: "Categorias", tabType: SearchFilter.category },
+  { label: "Usuários", tabType: SearchFilter.users },
+  { label: "Listas", tabType: SearchFilter.lists },
+];
 
 export const SearchScreen = () => {
+  const [search, setSearch] = useState<string>();
+  const [filter, setFilter] = useState<SearchFilter>(SearchFilter.title);
+
+  const TabComponent = useMemo(() => {
+    if (!search) {
+      return <></>;
+    }
+
+    return (
+      <SearchTabs
+        tabSelected={filter}
+        tabs={options}
+        handleChange={setFilter}
+        sx={{
+          mt: "24px",
+        }}
+      />
+    );
+  }, [search, filter, setFilter]);
+
   return (
     <Box>
-      <SearchHeader />
-      <SearchResultList />
+      <SearchHeader handleSearch={setSearch} search={search} />
+      <Box sx={{ mx: "24px" }}>
+        {TabComponent}
+        <SearchResultList />
+      </Box>
     </Box>
   );
 };
